Fixed Bug 29: Generated bindind objects have the following issue:
In a hierarchy where a variable size array is a member of a child, the
corresponding dynamic counter in the parent or higher (root actually) complex
type is not re-marshalled after children were marshalled.

This is due to the fact that there is no mechanism to keep track of dynamic
counters in static (ie generated) bindind classes.